(TK-192) Address PR feedback
Flesh out README with more examples and to list out all possible
values for the `type` parameter. Improve validation of `type`
parameter so that it only checks the parameter, and if it exists,
raises an exception if the specified type does not match the type
of the value passed in. Add tests for validation of the `type`
parameter.