Handle stringified numerics
Previously, the type would check the type of value being passed in the DSL
and would use that type when setting the value in the Hocon config file
(i.e. a string value would be a string in the config file, and an array
would be an array value). The problem is that versions of Puppet < 4.0.0
stringify all values passed in the DSL, so it was impossible to pass a true
numeric value via the DSL (since it would be stringified).

This commit updates the type validation to account for this and implements
a munge block to convert the value type to a numeric. A test for this case
has also been added.

squash