Determine whether a string is an integer or a float
Previously, there was no way to determine whether a string value was intended
to be an integer or a float.  This commit first attempts to cast a string value
as an integer. Failing that, an attempt to cast the string to a float is attempted.
If that fails, and the `type` property is set to 'number', an error is raised.

Tests were added to specifically test each case.