Add spec tests for the type
Previously, there were no spec tests to determine whether type validation and
munging was successful.  This commit adds them.  One test is commented out
because it is failing - because :array_matching is set to :all, EVERY
string/array/boolean/number value comes wrapped in an array. The validation
doesn't account for this currently, and a later commit will fix it.