Implement hocon::file convenience type
The hocon_setting type allows users to describe individual pieces of
configuration that they need, but doesn't let them step back and
describe at a higher level that a wholistic configuration file is what
they care about.

It is technically desirable not to use templates for managing
configuration, as it is impossible to be additive about configuration
with templates. However, templates make a lot of sense for defining a
baseline from which to start and be additive about. See
https://github.com/TERC/puppet-xmlfile for another example of this
desire.

This commit introduces a new defined type that allows users to build a
whole-file template representing a base configuration to deploy, but
does not have any of the brittleness of a template.

The defined type first declares the file resource but sets
replace=false. When Puppet runs it will create the file and populate it
with the template content.

Next, a function is added which parses the template content and creates
a hocon_setting resource for each element found in the template. These
resources are added to the catalog.

Because a template was used to create the file, formatting and comments
can be preserved on creation. Because Puppet will never overwrite an
existing file though, local edits or hocon_setting resources that add
additional configs will be work correctly.

Because a function parses the template and adds hocon_setting resources
to the catalog for each setting, Puppet will continue to enforce that
desired configuration is preserved, even though the file resource will
never replace the file.

In general the new hocon::file defined type allows users to manage Hocon
configuration files by focusing on the file instead of individual
settings inside it. Similar to the recurse=true file directory pattern,
users define a whole naturally in a wysiwyg manner, but aren't prevented
from adding in additional parameters later.