Remove unnecessary hocon require, add write_conf method
Prior to this commit we had a separate config_saver module
for saving the hocon config to disk. This module was requiring
the hocon gem without Puppet feature protection. This was causing
requirement issues on the Puppet Server side, where the hocon gem
is not typically installed.
This commit moves the code from config saver into the main hocon
provider, as the method write_conf. This removes the unnecessary
require, and cleans up the unnecessary config_saver module.