(HC-46) Confine provider based on hocon feature
Previously, we guarded requiring the hocon gem based on the feature, but
we never re-evaluated the feature, so you couldn't use the hocon module
in the same run that the hocon gem was installed.

This commit moves the requiring logic into the feature definition, and
confines the provider based on the hocon feature.